fix: rename Dialog2 -> Dialog
- tweak animations
- export Dialog.Title and Dialog.Description
